Hotel view touch ups.
A small quick fix to position the reception drape in its correct place and added the fix to the background image to repeat it properly. 
I also added a color switch for the widgets texts from black to white. 
All edits are as close as I could get it to being as accurate to Official Habbo as possible.