Update index.scss
Fixed the sized HC icon located in various places within the UI.

The image illustrates the follow -
Before - Left
After - Right

https://i.ibb.co/nPy5fJj/hcicon.png

This update will break the currency view in the purse, we will need to update that. But is the only solution. So fixing the currency view in the purse is necessary following this push.