Updated Pet List and Hotel View images.
Removed custom pets, added missing official pets and is 100% in sync with Morningstar database.

Updated the hotel view to be a real flash layout of the hotel view not the default one from the nitro's base UI.
Should make it feel more like the real deal upon installation.

Updated with the most current hotel view images being used by Habbo as of May.